fix: uniffi test bindings, fix warnings, cleanup dead code
